Contains code for working parsed stream.
Download, unzip, and follow instructions in readMe to run the website. The scroll box on the website should not contain any tweets. It will simply display a message saying "This is the area that the stream of tweets will be output to the screen."

This allows us to make changes to the website without actually requesting for tweets each time, which can cause the rate limit to exceed.